Refactoring: Move the Inner cache to common/base_cache module
