+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+    "rust-analyzer.cargo.features": ["runtime-tokio"],
     "cSpell.words": [
         "CLFU",
         "Deque",
@@ -9,6 +10,7 @@
         "Moka",
         "Ristretto",
         "Tatsuya",
+        "actix",
         "ahash",
         "benmanes",
         "clippy",
@@ -18,6 +20,7 @@
         "mpsc",
         "nanos",
         "nocapture",
+        "runtimes",
         "rustdoc",
         "rustfmt",
         "semver",
